[Feature][ClassExtendingExclusiveNamespaceRule] Simplify conditions, rename rule
