+++ packages/phpstan-rules/tests/Rules/ClassExtendingExclusiveNamespaceRule/Fixture/App/Component/PriceEngine/PriceProviderInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Symplify\PHPStanRules\Tests\Rules\ClassExtendingExclusiveNamespaceRule\Fixture\App\Component\PriceEngine;
+
+interface PriceProviderInterface
+{
+
+}
