+++ packages/phpstan-rules/tests/Rules/ClassExtendingExclusiveNamespaceRule/Fixture/App/Component/PriceEngineImpl/SkipDistributorProductProviderInAuthorizedNamespace.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Symplify\PHPStanRules\Tests\Rules\ClassExtendingExclusiveNamespaceRule\Fixture\App\Component\PriceEngineImpl;
+
+use Symplify\PHPStanRules\Tests\Rules\ClassExtendingExclusiveNamespaceRule\Fixture\App\Component\PriceEngine\ProductProviderInterface;
+
+class SkipDistributorProductProviderInAuthorizedNamespace implements ProductProviderInterface
+{
+
+}
