+++ packages/phpstan-rules/tests/Rules/ClassExtendingExclusiveNamespaceRule/Fixture/App/Model/Customer/CustomerRequestModelInUnauthorizedNamespace.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Symplify\PHPStanRules\Tests\Rules\ClassExtendingExclusiveNamespaceRule\Fixture\App\Model\Customer;
+
+use Symplify\PHPStanRules\Tests\Rules\ClassExtendingExclusiveNamespaceRule\Source\App\Component\ParamConverter\RequestInterface;
+
+class CustomerRequestModelInUnauthorizedNamespace implements RequestInterface
+{
+
+}
