+++ packages/phpstan-rules/tests/Rules/ClassExtendingExclusiveNamespaceRule/Fixture/App/Services/SkipExceptionInAuthorizedNamespace.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Symplify\PHPStanRules\Tests\Rules\ClassExtendingExclusiveNamespaceRule\Fixture\App\Services;
+
+use Exception;
+
+class SkipExceptionInAuthorizedNamespace extends Exception
+{
+}
