+++ packages/phpstan-rules/tests/Rules/ClassExtendingExclusiveNamespaceRule/Source/App/Component/ParamConverter/RequestInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Symplify\PHPStanRules\Tests\Rules\ClassExtendingExclusiveNamespaceRule\Source\App\Component\ParamConverter;
+
+interface RequestInterface
+{
+
+}
