+++ packages/phpstan-rules/tests/Rules/ClassExtendingExclusiveNamespaceRule/Source/Symfony/Component/Console/Command/Command.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Symplify\PHPStanRules\Tests\Rules\ClassExtendingExclusiveNamespaceRule\Source\Symfony\Component\Console\Command;
+
+class Command
+{
+
+}
