+++ packages/phpstan-rules/tests/Rules/ClassExtendingExclusiveNamespaceRule/Source/Symfony/Component/Form/FormTypeInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Symplify\PHPStanRules\Tests\Rules\ClassExtendingExclusiveNamespaceRule\Source\Symfony\Component\Form;
+
+interface FormTypeInterface
+{
+
+}
