--- packages/phpstan-rules/tests/Rules/ClassNamespaceGuardRule/Fixture/App/Command/SkipCommandInAuthorizedNamespace.php
@@ -1,11 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Symplify\PHPStanRules\Tests\Rules\ClassNamespaceGuardRule\Fixture\App\Command;
-
-use Symplify\PHPStanRules\Tests\Rules\ClassNamespaceGuardRule\Source\Symfony\Component\Console\Command\Command;
-
-class SkipCommandInAuthorizedNamespace extends Command
-{
-}
