--- packages/phpstan-rules/tests/Rules/ClassNamespaceGuardRule/Fixture/App/CommandInUnauthorizedNamespace.php
@@ -1,11 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Symplify\PHPStanRules\Tests\Rules\ClassNamespaceGuardRule\Fixture\App;
-
-use Symplify\PHPStanRules\Tests\Rules\ClassNamespaceGuardRule\Source\Symfony\Component\Console\Command\Command;
-
-class CommandInUnauthorizedNamespace extends Command
-{
-}
