--- packages/phpstan-rules/tests/Rules/ClassNamespaceGuardRule/Fixture/App/Component/PriceEngine/SkipFallbackPriceProviderInAuthorizedNamespace.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Symplify\PHPStanRules\Tests\Rules\ClassNamespaceGuardRule\Fixture\App\Component\PriceEngine;
-
-class SkipFallbackPriceProviderInAuthorizedNamespace implements PriceProviderInterface
-{
-
-}
