--- packages/phpstan-rules/tests/Rules/ClassNamespaceGuardRule/Fixture/App/Component/PriceEngineImpl/SkipCustomerProductProviderInAuthorizedNamespace.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Symplify\PHPStanRules\Tests\Rules\ClassNamespaceGuardRule\Fixture\App\Component\PriceEngineImpl;
-
-use Symplify\PHPStanRules\Tests\Rules\ClassNamespaceGuardRule\Fixture\App\Component\PriceEngine\ProductProviderInterface;
-
-class SkipCustomerProductProviderInAuthorizedNamespace implements ProductProviderInterface
-{
-
-}
