--- packages/phpstan-rules/tests/Rules/ClassNamespaceGuardRule/Fixture/App/Component/PriceEngineImpl/SkipDistributorPriceProviderInAuthorizedNamespace.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Symplify\PHPStanRules\Tests\Rules\ClassNamespaceGuardRule\Fixture\App\Component\PriceEngineImpl;
-
-use Symplify\PHPStanRules\Tests\Rules\ClassNamespaceGuardRule\Fixture\App\Component\PriceEngine\PriceProviderInterface;
-
-class SkipDistributorPriceProviderInAuthorizedNamespace implements PriceProviderInterface
-{
-
-}
