--- packages/phpstan-rules/tests/Rules/ClassNamespaceGuardRule/Fixture/App/Component/PriceProviderInUnauthorizedNamespace.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Symplify\PHPStanRules\Tests\Rules\ClassNamespaceGuardRule\Fixture\App\Component;
-
-use Symplify\PHPStanRules\Tests\Rules\ClassNamespaceGuardRule\Fixture\App\Component\PriceEngine\PriceProviderInterface;
-
-class PriceProviderInUnauthorizedNamespace implements PriceProviderInterface
-{
-
-}
