--- packages/phpstan-rules/tests/Rules/ClassNamespaceGuardRule/Fixture/App/ExceptionInUnauthorizedNamespace.php
@@ -1,11 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Symplify\PHPStanRules\Tests\Rules\ClassNamespaceGuardRule\Fixture\App;
-
-use Exception;
-
-class ExceptionInUnauthorizedNamespace extends Exception
-{
-}
