--- packages/phpstan-rules/tests/Rules/ClassNamespaceGuardRule/Fixture/App/Form/SkipFormInAuthorizedNamespace.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Symplify\PHPStanRules\Tests\Rules\ClassNamespaceGuardRule\Fixture\App\Form;
-
-use Symplify\PHPStanRules\Tests\Rules\ClassNamespaceGuardRule\Source\Symfony\Component\Form\FormTypeInterface;
-
-class SkipFormInAuthorizedNamespace implements FormTypeInterface
-{
-}
