--- packages/phpstan-rules/tests/Rules/ClassNamespaceGuardRule/Fixture/App/FormInUnauthorizedNamespace.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Symplify\PHPStanRules\Tests\Rules\ClassNamespaceGuardRule\Fixture\App;
-
-use Symplify\PHPStanRules\Tests\Rules\ClassNamespaceGuardRule\Source\Symfony\Component\Form\FormTypeInterface;
-
-class FormInUnauthorizedNamespace implements FormTypeInterface
-{
-}
