--- packages/phpstan-rules/tests/Rules/ClassNamespaceGuardRule/Fixture/App/Model/Customer/Request/SkipCustomerRequestModelInAuthorizedNamespace.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Symplify\PHPStanRules\Tests\Rules\ClassNamespaceGuardRule\Fixture\App\Model\Customer\Request;
-
-use Symplify\PHPStanRules\Tests\Rules\ClassNamespaceGuardRule\Source\App\Component\ParamConverter\RequestInterface;
-
-class SkipCustomerRequestModelInAuthorizedNamespace implements RequestInterface
-{
-
-}
