--- packages/phpstan-rules/tests/Rules/ClassNamespaceGuardRule/Fixture/App/Model/Order/OrderRequestModelInUnauthorizedNamespace.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Symplify\PHPStanRules\Tests\Rules\ClassNamespaceGuardRule\Fixture\App\Model\Order;
-
-use Symplify\PHPStanRules\Tests\Rules\ClassNamespaceGuardRule\Source\App\Component\ParamConverter\RequestInterface;
-
-class OrderRequestModelInUnauthorizedNamespace implements RequestInterface
-{
-
-}
