--- packages/phpstan-rules/tests/Rules/ClassNamespaceGuardRule/Fixture/App/Services/SkipExceptionInAuthorizedNamespace.php
@@ -1,11 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Symplify\PHPStanRules\Tests\Rules\ClassNamespaceGuardRule\Fixture\App\Services;
-
-use Exception;
-
-class SkipExceptionInAuthorizedNamespace extends Exception
-{
-}
