--- packages/phpstan-rules/tests/Rules/ClassNamespaceGuardRule/Source/App/Component/ParamConverter/RequestInterface.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Symplify\PHPStanRules\Tests\Rules\ClassNamespaceGuardRule\Source\App\Component\ParamConverter;
-
-interface RequestInterface
-{
-
-}
