--- packages/phpstan-rules/tests/Rules/ClassNamespaceGuardRule/Source/Symfony/Component/Console/Command/Command.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Symplify\PHPStanRules\Tests\Rules\ClassNamespaceGuardRule\Source\Symfony\Component\Console\Command;
-
-class Command
-{
-
-}
