--- packages/phpstan-rules/tests/Rules/ClassNamespaceGuardRule/Source/Symfony/Component/Form/FormTypeInterface.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Symplify\PHPStanRules\Tests\Rules\ClassNamespaceGuardRule\Source\Symfony\Component\Form;
-
-interface FormTypeInterface
-{
-
-}
